Combustion chamber Main(see description))
Ok so there were a bunch of changes in the CC model. Inputs are not completely sanitized but I've started working on it. If it quits the program it means values coming from injector are no good, and the convergence ratio will either be too big or too small. As long as CC gets reasonable values (d0 bigger than 1.35*10^-5, smaller than 0.5*10**-3, injection velocity greater than 15m/s) it should be okay. I'm going to start working on the default values now