CC + igniters changes - read description
Igniters and CC values put on default category (I think i did not miss anything)
filled data category as well
I've noticed a problem with the value that is being inputted into the CC for the droplet initial diameter. For now do not change the inputs of the CC and it should run (it does on my pc), monday we can discuss the problem that i found.